pdf-gpt/requirements.txt
docarray==0.41.0
gradio==5.25.0
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.51
langchain_openai==0.3.12
pdfplumber==0.11.5
Pillow==11.1.0
PyMuPDF==1.25.5
python-dotenv==1.1.0
redis==5.2.1
